Fixed two tests in python client.
test_type_checking: the client is creating a new key, and thus returns SUCCESS_KEY_CREATED instead of SUCCESS.

test_looping: the purpose behind the __iter__ function for LazySysAdmin is to iterate through the unique keys (regardless of subkeys) that have been set, at the namespace request. If the keys being set are the following:
	network.dhcp.interfaces
	network.breakfast
at the top namespace, this means there is 1 key to iterate through (network). At the 'network' namespace, there are 2 keys to iterate through (dhcp, breakfast). This means that there are only 4 keys to iterate through in the test_looping test (network, my, default, loops).